disable AD if docker not compiled in
We'll need to register the DockerListener through a catalog
instead of hardcoding its init. Will be done when adding the
KubeletListener
